+++ BeyondTrust/tests/conftest.py
@@ -99,3 +99,60 @@ def error_response_xml():
 <error xmlns="http://www.beyondtrust.com/sra/namespaces/API/reporting" code="authentication_failed">
 Authentication failed. Please check your credentials.
 </error>"""
+
+
+@pytest.fixture
+def vault_activity_list():
+    return b"""<?xml version="1.0" encoding="UTF-8"?>
+        <vault_account_activity_list xmlns="http://www.beyondtrust.com/sra/namespaces/API/reporting">
+           <start_time timestamp="986204380">2001-04-02T09:39:40+00:00</start_time>
+           <end_time timestamp="1775127348">2026-04-02T10:55:48+00:00</end_time>
+           <vault_account_activity timestamp="1773671926" event_type="Account Created" account="1">
+              <performed_by id="2" type="User">John Doe</performed_by>
+              <data></data>
+           </vault_account_activity>
+           <vault_account_activity timestamp="1773672020" event_type="Password Changed" account="1">
+              <performed_by id="2" type="User">John Doe</performed_by>
+              <data>Manually Edited</data>
+           </vault_account_activity>
+           <vault_account_activity timestamp="1773672107" event_type="Account Deleted" account="1">
+              <performed_by id="2" type="User">John Doe</performed_by>
+              <data></data>
+           </vault_account_activity>
+        </vault_account_activity_list>"""
+
+
+@pytest.fixture
+def team_xml():
+    return b"""<?xml version="1.0" encoding="UTF-8"?>
+<team_activity_list xmlns="http://www.beyondtrust.com/sra/namespaces/API/reporting">
+    <start_time timestamp="1772323200">2026-03-01T00:00:00+00:00</start_time>
+    <end_time timestamp="1773679977">2026-03-16T16:52:57+00:00</end_time>
+    <team_activity name="Integration" id="1">
+        <logged_in_representatives>
+        </logged_in_representatives>
+        <events>
+            <event timestamp="1773679963" event_type="Conference Member Added">
+                <performed_by gsnumber="15" type="representative">John Doe</performed_by>
+                <data>
+                    <value name="os" value="Unknown"/>
+                    <value name="private_ip" value="Unknown"/>
+                    <value name="private_name" value="John Doe"/>
+                    <value name="public_ip" value="4.3.2.1:41286"/>
+                    <value name="public_name" value="John Doe"/>
+                    <value name="support_teams" value="Integration"/>
+                    <value name="user_id" value="2"/>
+                </data>
+            </event>
+            <event timestamp="1773679963" event_type="Conference Member State Changed">
+                <performed_by gsnumber="15" type="representative">John Doe</performed_by>
+                <data>
+                    <value name="os" value="Unknown"/>
+                    <value name="private_ip" value="Unknown"/>
+                    <value name="public_ip" value="4.3.2.1:41286"/>
+                    <value name="state" value="connected"/>
+                </data>
+            </event>
+        </events>
+    </team_activity>
+</team_activity_list>"""