+++ AWS/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 2026-04-28 - 1.35.3
+
+### Changed
+
+- Remove useless file
+
 ## 2026-04-20 - 1.35.2
 
 ### Changed