+++ MicrosoftEntraID/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 2026-04-21 - 2.10.20
+
+### Changed
+
+- Update the sekoia-automation-sdk dependency to version 1.22.5
+
 ## 2026-03-29 - 2.10.19
 
 ### Changed